requirements.txt
fastapi
uvicorn
pandas
transformers
torch
spacy
pytest
python-dotenv
loguru
httpx
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0.tar.gz
